+++ server/controllers/todoController.js
@@ -26,6 +26,7 @@ async function getTodo (req, res, id) {
 }
 
 async function addTodo (req, res) {
+  console.log(req)
   try {
     const newtodo = {
       title: req.body.title,
